Move age restriction into attribute package
