+++ src/main/java/net/dv8tion/jda/api/audit/AuditLogKey.java
@@ -244,7 +244,7 @@ public enum AuditLogKey
     CHANNEL_USER_LIMIT("user_limit"),
 
     /**
-     * Change of the {@link net.dv8tion.jda.api.entities.channel.IAgeRestrictedChannel#isNSFW() IAgeRestrictedChannel.isNSFW()} value.
+     * Change of the {@link net.dv8tion.jda.api.entities.channel.attribute.IAgeRestrictedChannel#isNSFW() IAgeRestrictedChannel.isNSFW()} value.
      *
      * <p>Expected type: <b>Boolean</b>
      */
