+++ src/main/java/net/dv8tion/jda/api/entities/StandardGuildMessageChannel.java
@@ -16,7 +16,7 @@
 
 package net.dv8tion.jda.api.entities;
 
-import net.dv8tion.jda.api.entities.channel.IAgeRestrictedChannel;
+import net.dv8tion.jda.api.entities.channel.attribute.IAgeRestrictedChannel;
 import net.dv8tion.jda.api.managers.channel.middleman.StandardGuildMessageChannelManager;
 import net.dv8tion.jda.api.requests.restaction.ChannelAction;
